+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "acceptNPCLeftClick: true / false" config option
 - Added optional "minlevel" and "maxlevel" arguments to mmobkill objective
 - Added new options 'inside/outside' for npcrange objective, support for multiple npcs and improved performance
+- Added new Event QuestCompassTargetChangeEvent that is triggered when a new CompassTarget is set. It is also possible to cancel it
 ### Changed
 - devbuilds always show notifications for new devbuilds, even when the user is not on a _DEV strategy
 - Items for HolographicDisplays are now defines in items.yml